main: Fix LPM4 sleep mode bug
This reduces current drain in sleep mode after the first button
press by ~80uA.

I had forgotten that the LPM4_EXIT macro clears the LPM4 bits on the
exit of the subroutine, rather than setting them. I knew I needed the
on_exit as the SR is popped on return from the interrupt. There
seems to be no macro for exiting into a lower power mode, so I'm
using the compiler intrinsic instead.